Capstone Project/requirements.txt
imbalanced_learn==0.10.1
openai==0.27.5
pandas==2.0.1
scikit_learn==1.2.2
google-cloud==0.34.0
google-cloud-storage==2.8.0
google-cloud-videointelligence==2.11.1
fsspec==2023.4.0
gcsfs==2023.4.0
